Add cider AppImage as overlay

+++ overlays/cider-appimage.nix
@@ -3,7 +3,10 @@ self: super: with super; {
     pname = "cider";
     version = "3.0.2";
 
-    src = ./cider-v3.0.2-linux-x64.AppImage;
+    src = builtins.path {
+      path = ./cider-v3.0.2-linux-x64.AppImage;
+      name = "cider-v${version}-linux-x64.AppImage";
+    };
 
     nativeBuildInputs = [ makeWrapper ];
 
@@ -16,14 +19,14 @@ self: super: with super; {
       in
       ''
         wrapProgram $out/bin/${pname} \
-           --add-flags "\''${NIXOS_OZONE_WL:+\''${WAYLAND_DISPLAY:+--ozone-platform-hint=auto --enable-features=WaylandWindowDecorations --enable-wayland-ime=true}}" \
+           --add-flags "--ozone-platform=wayland --enable-features=UseOzonePlatform,WaylandWindowDecorations" \
            --add-flags "--no-sandbox --disable-gpu-sandbox"
 
-        install -m 444 -D ${contents}/Cider.desktop $out/share/applications/${pname}.desktop
-        substituteInPlace $out/share/applications/${pname}.desktop \
-          --replace-warn 'Exec=Cider' 'Exec=${pname}'
-        install -Dm444 ${contents}/usr/share/icons/hicolor/256x256/cider-linux----arch-------version---.png \
-                       $out/share/icons/hicolor/256x256/apps/cider.png
+        if [ -f ${contents}/Cider.desktop ]; then
+          install -m 444 -D ${contents}/Cider.desktop $out/share/applications/${pname}.desktop
+          substituteInPlace $out/share/applications/${pname}.desktop \
+            --replace-warn 'Exec=Cider' 'Exec=${pname}'
+        fi
       '';
 
     meta = with lib; {
